requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
neo4j>=5.0.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
numpy~=1.26.4
pandas>=2.0.0

# API and Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
aiofiles>=23.0.0

# Streamlit UI
streamlit>=1.28.0
plotly>=5.15.0
streamlit-aggrid

# HTTP client for external services
httpx>=0.25.0
requests>=2.31.0
aiohttp>=3.8.0

# External LLM providers
openai>=1.0.0
google-generativeai>=0.8.0

# File upload support for FastAPI
python-multipart>=0.0.6

# Configuration and Environment
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Dependency Injection
dependency-injector>=4.41.0

# Database and Caching
redis>=4.5.0
asyncio-pool>=0.5.0

# File Processing
pypdf>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
python-magic>=0.4.27

# Text Processing
beautifulsoup4>=4.12.0
markdown>=3.5.0
tiktoken>=0.5.0

# Monitoring and Observability
prometheus-client>=0.17.0
structlog>=23.0.0
sentry-sdk>=1.32.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
httpx>=0.24.0
factory-boy>=3.3.0

# Development
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Type hints
types-requests>=2.31.0
types-redis>=4.6.0

# Utilities
click>=8.1.0
rich>=13.5.0
tqdm>=4.65.0
python-dateutil>=2.8.2
